requirements.txt
torch
torchvision
numpy
pillow
fastapi
uvicorn
python-multipart
diffusers
transformers
accelerate
safetensors
pydantic-settings
peft